requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap5==2025.4
diff-match-patch==20241021
Django==5.2
django-crispy-forms==2.3
django-image-optimizer==1.0.3
django-import-export==4.3.7
idna==3.10
pillow==11.1.0
python-resize-image==1.1.20
requests==2.32.3
sqlparse==0.5.3
tablib==3.8.0
tinify==1.7.0
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
